refactor: 거리 계산 클래스도 calculator 패키지 내부로 이동
